Organize all profiles in subdirectories
All profiles should be put in /profiles/APPLICATION/FILENAME.nix.

I noticed that there can be multiple profiles for a single application.
It is tedious to rename all imports every time I add a secondary profile
of the same application, so I will put every profile in a subdirectory
and make default.nix the default profile of an application.